requirements.txt
awkward==2.6.5
comet_ml==3.49.6
lightning==2.5.0.post0
pytorch_lightning==2.5.0.post0
matplotlib==3.9.1
numpy==1.26.4
pandas==2.2.2
pyyaml_include==2.1
scikit_learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
torch==2.6.0+cu126
tqdm==4.66.4
uproot==5.3.12